split tests from lib.rs
